Les compétences sont dans le personnage
