+++ core/src/main/java/io/github/algorys/agshmne/character/skills/factory/Fouiller.java
@@ -1,6 +1,6 @@
-package io.github.algorys.agshmne.skills.factory;
+package io.github.algorys.agshmne.character.skills.factory;
 
-import io.github.algorys.agshmne.skills.ISkillFactory;
+import io.github.algorys.agshmne.character.skills.ISkillFactory;
 
 public class Fouiller implements ISkillFactory {
 	private int level;
