+++ swing/src/main/java/io/github/algorys/agshmne/start/JFicheNav.java
@@ -18,9 +18,9 @@ import javax.swing.JFrame;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import io.github.algorys.agshmne.character.Character;
+import io.github.algorys.agshmne.character.skills.SkillType;
 import io.github.algorys.agshmne.game.JGame;
 import io.github.algorys.agshmne.region.Region;
-import io.github.algorys.agshmne.skills.SkillType;
 import io.github.algorys.agshmne.tile.JTile;
 
 @SuppressWarnings("serial")
